src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/anchor-content-projected-expected.txt

CONSOLE ERROR: line 27: Uncaught TypeError: host.createShadowRoot is not a function
Links should be clickable even when their content is projected into them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught TypeError: host.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

Example link
